Allow empty INDIVIDUAL_ID and LIBRARY_ID columns in input file
Process createSE() had required that the INDIVIDUAL_ID and LIBRARY_ID columns of the input file be nonblank (i.e., they should be set to the same as SAMPLE_ID if library:sample:individual is 1:1:1).  However, this is a holdover from earlier versions of the pipeline that included the variant calling steps, and so to prevent accidental crashes, process createSE() has been edited to remove the requirement for these columns to be nonblank.

NOTE: for backwards compatibility, and because the variant calling functionality may be reintroduced in the future, these two columns are still required to be present in the input file, even if blank.

README.md has also been edited to remove the requirement that these columns be nonblank.