Fix issue introduced by previous commit
The intent of the previous commit was to allow the INDIVIDUAL_ID and LIBRARY_ID columns of the input file to remain empty; however, the code included a "magic number", i.e., the number of aggregation variables was assumed to be 3 (INDIVIDUAL_ID, SAMPLE_ID, and LIBRARY_ID).  As a result, when aggregation was performed using only SAMPLE_ID, and the next step was executed to remove the (redundant) columns containing the aggregation variables, three columns (SAMPLE_ID, INDIVIDUAL_ID, and SAMPLE_ID again) were removed instead of only the first (i.e., redundant) copy of SAMPLE_ID.

To prevent this problem in the future, the code has been refactored so that the number of aggregation variables is counted rather than assumed.  A comment was also generalized accordingly.