Upgrade to use any Ensembl version
The main purpose of this commit is to enable users to specify any Ensembl version in the .config file, decoupling RNA_Seq.nf from locally maintained Ensembl GTF and BED files and RSEM reference files.  These files had previously been generated using standalone scripts, but as they are relatively small (GTF and BED < 100MB, RSEM reference < 2GB) and can be generated very quickly, it makes sense to simply download/process/generate them on the fly for a given Ensembl version.  The only new dependencies that this commit introduces are on the availability of the Ensembl FTP site (for the GTF file) and the UCSC web site (for the chromAlias table and for command-line tools to convert the GTF file to a BED file).

The changes in this commit include:

* New process `generateGTF`, which does the following:
  - Retrieves a GTF file from Ensembl for a given species, genome assembly, and Ensembl build.
  - Retrieves the chromAlias table from UCSC for the corresponding UCSC build.
  - Uses the chromAlias table to convert the sequence (chromosome) names in the GTF from Ensembl nomenclature to UCSC nomenclature.  (Note: the external script that had previously carried out this step was written with the incorrect assumption that UCSC sequence names are simply generated by prepending "chr" to the Ensembl sequence name; in human and mouse, this assumption only holds for chromosomes 1-22, X, and Y, but not for the mitochondrial chromosome or any random contigs.)
  - Limits the GTF file to the sequences present in the FASTA reference specified in the .config file.
  - Stores the result as a gzipped file.

* New process `generateBED`, which uses the UCSC command-line tools `gtfToGenePred` and `genePredToBed` to convert the GTF file to a BED file.

* New process `runRSEMprepareReference`, which uses the FASTA reference and the Ensembl GTF file to build an RSEM reference.

* Renaming of processes and channels using camelCase for consistency and easier readability/maintainability.  (Also corrected one channel name that mistakenly referred to the "read_GC" step as "QC".)

* Edits to RNA_Seq_template.config to remove the dependencies on the local Ensembl GTF and BED files and RSEM reference, allow the specification of a reference subset (e.g., "base_random"), reorganize URLs, and provide SGE settings for new processes.

* Updates to documentation in README.md, including more details about the fields within `params.genome`.