Fix problem in process `createSE`, update README
This commit fixes an issue in process `createSE`, namely, that it assumes that FASTQ filenames will contain "_R1" or "_R2".  Although this is the convention for filenames generated by Illumina software, it is not universal; therefore, this requirement has been loosened so that the filename prefix (before .fastq* or .fq*) must end in "1" or "2".

The README documentation has also been updated to clarify and correct some points about the contents of the input file.