Fix process generateGTF
This commit fixes a bug in process `generateGTF`.  The command to write the GTF includes a step to limit the output to those sequences that are represented in the input FASTA file, using grep with a regex built from the sequence names, anchored to the start of the line.  This edit adds a terminal tab character to the regex (and changes the grep command to Perl-style mode) so that it will match on the entire value in the first column rather than just the initial part of the value (e.g., regex "chrX\t" will match only lines with first-column value "chrX" and not lines with first-column values like "chrX_gl877877_fix").