Accommodate Ensembl genome reference/sequence IDs
This commit is motivated by a request to modify it to be able to use a reference genome FASTA file from Ensembl and to use Ensembl sequence identifiers (e.g., "1", "X", "MT") rather than UCSC sequence identifiers (e.g., "chr1", "chrX", "chrM").  Accordingly, the config file and README have been edited so that deleting params.genome.ucsc or setting it to the empty string "" will instruct process generateGTF to assume that the FASTA file contains Ensembl sequence identifiers and to skip the step of translating the sequence identifiers in the GTF file from Ensembl to UCSC format.

This commit also includes edits to the config file to move a few parameters into the "Run-specific variables" section, update some software to more recent versions, and remove references to some software packages that are not currently used.