Update runSTARgenomeGenerate to accommodate STAR 2.79a
The previous commit upgraded the default version of STAR from 2.6.0c to 2.7.9a, and the newer version creates the file Log.out in the path specified by the --genomeDir argument, rather than in the current working directory.  As a result, the dedicated output channel for Log.out has been deprecated.  An additional command has also been added at the end of the process to move Log.out into the genome directory if it exists (i.e., if the user chose to use an earlier version of STAR for some reason).