Fix simple bug in runSTARgenomeGenerate()
This commit fixes a simple error in process `runSTARgenomeGenerate` that was causing it to crash if the user specified STAR version 2.6.0c.  It was trying to copy the file `Log.out` to the genome input directory used by process `runSTAR1pass` rather than by the genome output directory used by process `runSTARgenomeGenerate`.