Fix another bug in runSTARgenomeGenerate
This commit fixes a trivial but frustrating bug introduced in runSTARgenomeGenerate in commit f4727873a7471c7c2a223ec8e2d67ea1d367967c.

At the end of the process, a conditional statement checks whether the file 'Log.out' is present in the current working directory, and if so, moves it to the output directory along with the other files.  However, the statement did not include any action to be performed if the file was not present, and so, in this case, it would return exit code 1, causing the process to fail (even if the genomeGenerate step worked correctly).  This case is now handled with the 'true' command to return exit code 0 if the file is not present.